+++ tests/adaptive_map/test_adaptive_map_entry.c
@@ -135,7 +135,7 @@ check_static_begin(adaptive_map_test_insert) {
     check(v->key, -1);
     int i = 0;
 
-    check(fill_n(&om, size / 2, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size / 2, i, &allocator), CHECK_PASS);
 
     i += (size / 2);
     ent = swap_entry(
@@ -163,7 +163,7 @@ check_static_begin(adaptive_map_test_insert) {
     check(v->key, i);
     ++i;
 
-    check(fill_n(&om, size - i, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size - (size_t)i, i, &allocator), CHECK_PASS);
 
     i = size;
     ent = swap_entry(
@@ -230,7 +230,7 @@ check_static_begin(adaptive_map_test_remove_key_value) {
     check(v->key, -1);
     int i = 0;
 
-    check(fill_n(&om, size / 2, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size / 2, i, &allocator), CHECK_PASS);
 
     i += (size / 2);
     ent = CCC_remove_key_value(
@@ -260,7 +260,7 @@ check_static_begin(adaptive_map_test_remove_key_value) {
     check(v->val, i);
     check(v->key, i);
 
-    check(fill_n(&om, size - i, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size - (size_t)i, i, &allocator), CHECK_PASS);
 
     i = size;
     ent = CCC_remove_key_value(
@@ -317,7 +317,7 @@ check_static_begin(adaptive_map_test_try_insert) {
     check(v->key, -1);
     int i = 0;
 
-    check(fill_n(&om, size / 2, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size / 2, i, &allocator), CHECK_PASS);
 
     i += (size / 2);
     ent = try_insert(&om, &(struct Val){.key = i, .val = i}.elem, &allocator);
@@ -335,7 +335,7 @@ check_static_begin(adaptive_map_test_try_insert) {
     check(v->key, i);
     ++i;
 
-    check(fill_n(&om, size - i, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size - (size_t)i, i, &allocator), CHECK_PASS);
 
     i = size;
     ent = try_insert(&om, &(struct Val){.key = i, .val = i}.elem, &allocator);
@@ -377,7 +377,7 @@ check_static_begin(adaptive_map_test_try_insert_with) {
     check(v->key, -1);
     int i = 0;
 
-    check(fill_n(&om, size / 2, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size / 2, i, &allocator), CHECK_PASS);
 
     i += (size / 2);
     ent = adaptive_map_try_insert_with(&om, i, &allocator, val(i));
@@ -395,7 +395,7 @@ check_static_begin(adaptive_map_test_try_insert_with) {
     check(v->key, i);
     ++i;
 
-    check(fill_n(&om, size - i, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size - (size_t)i, i, &allocator), CHECK_PASS);
 
     i = size;
     ent = adaptive_map_try_insert_with(&om, i, &allocator, val(i));
@@ -442,7 +442,7 @@ check_static_begin(adaptive_map_test_insert_or_assign) {
     check(v->key, -1);
     int i = 0;
 
-    check(fill_n(&om, size / 2, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size / 2, i, &allocator), CHECK_PASS);
 
     i += (size / 2);
     ent = insert_or_assign(
@@ -463,7 +463,7 @@ check_static_begin(adaptive_map_test_insert_or_assign) {
     check(v->key, i);
     ++i;
 
-    check(fill_n(&om, size - i, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size - (size_t)i, i, &allocator), CHECK_PASS);
 
     i = size;
     ent = insert_or_assign(
@@ -511,7 +511,7 @@ check_static_begin(adaptive_map_test_insert_or_assign_with) {
     check(v->key, -1);
     int i = 0;
 
-    check(fill_n(&om, size / 2, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size / 2, i, &allocator), CHECK_PASS);
 
     i += (size / 2);
     ent = adaptive_map_insert_or_assign_with(&om, i, &allocator, val(i));
@@ -528,7 +528,7 @@ check_static_begin(adaptive_map_test_insert_or_assign_with) {
     check(v->key, i);
     ++i;
 
-    check(fill_n(&om, size - i, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size - (size_t)i, i, &allocator), CHECK_PASS);
 
     i = size;
     ent = adaptive_map_insert_or_assign_with(&om, i, &allocator, val(i));
@@ -581,7 +581,7 @@ check_static_begin(adaptive_map_test_entry_and_modify) {
     check(v->val, 0);
     int i = 0;
 
-    check(fill_n(&om, size / 2, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size / 2, i, &allocator), CHECK_PASS);
 
     i += (size / 2);
     ent = entry(&om, &i);
@@ -600,7 +600,7 @@ check_static_begin(adaptive_map_test_entry_and_modify) {
     check(v->key, i);
     ++i;
 
-    check(fill_n(&om, size - i, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size - (size_t)i, i, &allocator), CHECK_PASS);
 
     i = size;
     ent = entry(&om, &i);
@@ -663,7 +663,7 @@ check_static_begin(adaptive_map_test_entry_and_context_modify) {
     check(v->val, 0);
     int i = 0;
 
-    check(fill_n(&om, size / 2, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size / 2, i, &allocator), CHECK_PASS);
 
     i += (size / 2);
     ent = entry(&om, &i);
@@ -694,7 +694,7 @@ check_static_begin(adaptive_map_test_entry_and_context_modify) {
     check(count(&om).count, i + 2);
     ++i;
 
-    check(fill_n(&om, size - i, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size - (size_t)i, i, &allocator), CHECK_PASS);
 
     i = size;
     ent = entry(&om, &i);
@@ -756,7 +756,7 @@ check_static_begin(adaptive_map_test_entry_and_modify_with) {
     check(count(&om).count, 1);
     int i = 0;
 
-    check(fill_n(&om, size / 2, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size / 2, i, &allocator), CHECK_PASS);
 
     i += (size / 2);
     ent = adaptive_map_entry_wrap(&om, &i);
@@ -775,7 +775,7 @@ check_static_begin(adaptive_map_test_entry_and_modify_with) {
     check(count(&om).count, i + 2);
     ++i;
 
-    check(fill_n(&om, size - i, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size - (size_t)i, i, &allocator), CHECK_PASS);
 
     i = size;
     ent = adaptive_map_entry_wrap(&om, &i);
@@ -825,7 +825,7 @@ check_static_begin(adaptive_map_test_or_insert) {
     check(count(&om).count, 1);
     int i = 0;
 
-    check(fill_n(&om, size / 2, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size / 2, i, &allocator), CHECK_PASS);
 
     i += (size / 2);
     v = or_insert(
@@ -849,7 +849,7 @@ check_static_begin(adaptive_map_test_or_insert) {
     check(count(&om).count, i + 2);
     ++i;
 
-    check(fill_n(&om, size - i, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size - (size_t)i, i, &allocator), CHECK_PASS);
 
     i = size;
     v = or_insert(
@@ -900,7 +900,7 @@ check_static_begin(adaptive_map_test_or_insert_with) {
     check(count(&om).count, 1);
     int i = 0;
 
-    check(fill_n(&om, size / 2, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size / 2, i, &allocator), CHECK_PASS);
 
     i += (size / 2);
     v = adaptive_map_or_insert_with(
@@ -920,7 +920,7 @@ check_static_begin(adaptive_map_test_or_insert_with) {
     check(count(&om).count, i + 2);
     ++i;
 
-    check(fill_n(&om, size - i, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size - (size_t)i, i, &allocator), CHECK_PASS);
 
     i = size;
     v = adaptive_map_or_insert_with(
@@ -971,7 +971,7 @@ check_static_begin(adaptive_map_test_insert_entry) {
     check(count(&om).count, 1);
     int i = 0;
 
-    check(fill_n(&om, size / 2, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size / 2, i, &allocator), CHECK_PASS);
 
     i += (size / 2);
     v = insert_entry(
@@ -995,7 +995,7 @@ check_static_begin(adaptive_map_test_insert_entry) {
     check(count(&om).count, i + 2);
     ++i;
 
-    check(fill_n(&om, size - i, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size - (size_t)i, i, &allocator), CHECK_PASS);
 
     i = size;
     v = insert_entry(
@@ -1046,7 +1046,7 @@ check_static_begin(adaptive_map_test_insert_entry_with) {
     check(count(&om).count, 1);
     int i = 0;
 
-    check(fill_n(&om, size / 2, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size / 2, i, &allocator), CHECK_PASS);
 
     i += (size / 2);
     v = adaptive_map_insert_entry_with(
@@ -1066,7 +1066,7 @@ check_static_begin(adaptive_map_test_insert_entry_with) {
     check(count(&om).count, i + 2);
     ++i;
 
-    check(fill_n(&om, size - i, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size - (size_t)i, i, &allocator), CHECK_PASS);
 
     i = size;
     v = adaptive_map_insert_entry_with(
@@ -1114,7 +1114,7 @@ check_static_begin(adaptive_map_test_remove_entry) {
     check(count(&om).count, 0);
     int i = 0;
 
-    check(fill_n(&om, size / 2, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size / 2, i, &allocator), CHECK_PASS);
 
     i += (size / 2);
     v = or_insert(
@@ -1134,7 +1134,7 @@ check_static_begin(adaptive_map_test_remove_entry) {
     check(occupied(e), true);
     check(count(&om).count, i);
 
-    check(fill_n(&om, size - i, i, &allocator), CHECK_PASS);
+    check(fill_n(&om, (size_t)size - (size_t)i, i, &allocator), CHECK_PASS);
 
     i = size;
     v = or_insert(
